refactor: TsunamiPlaybackSelectionState モデルを専用ファイルへ分離
これまでのコミットで削除漏れだった stale generated file も併せて削除する。
