+++ app/lib/feature/tsunami/data/model/tsunami_playback_selection_state.freezed.dart
@@ -3,7 +3,7 @@
 // ignore_for_file: type=lint, type=warning, deprecated_member_use, deprecated_member_use_from_same_package
 // ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
 
-part of 'tsunami_playback_selection_notifier.dart';
+part of 'tsunami_playback_selection_state.dart';
 
 // **************************************************************************
 // FreezedGenerator
